pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "biomni-mcp-server"
version = "0.1.0"
description = "MCP server for bioinformatics CLI tools on Databricks"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "databricks-sdk>=0.20.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "main.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
